+++ plugin-code/src/test/java/com/trivago/rta/features/FeatureFileConverterTest.java
@@ -78,14 +78,14 @@ public class FeatureFileConverterTest {
         when(fileIO.readContentFromFile("TEST_PATH")).thenReturn("TEST_CONTENT");
 
         List<SingleScenario> scenarioList = new ArrayList<>();
-        SingleScenario singleScenario = new SingleScenario("feature", "","featureDescription", "name", "scenarioDescription", new ArrayList<>(), new ArrayList<>());
+        SingleScenario singleScenario = new SingleScenario("feature", "","","featureDescription", "name", "scenarioDescription", new ArrayList<>(), new ArrayList<>());
         scenarioList.add(singleScenario);
-        when(gherkinDocumentParser.getSingleScenariosFromFeature("TEST_CONTENT", null, null, null)).thenReturn(scenarioList);
+        when(gherkinDocumentParser.getSingleScenariosFromFeature("TEST_CONTENT", "TEST_PATH",null, null, null)).thenReturn(scenarioList);
 
         String featureFileContent = "test";
         when(featureFileContentRenderer.getRenderedFeatureFileContent(singleScenario)).thenReturn(featureFileContent);
 
-        when(runnerFileContentRenderer.getRenderedRunnerFileContent(any(SingleScenarioRunner.class))).thenReturn("RUNNER_CONTENT");
+        when(runnerFileContentRenderer.getRenderedRunnerFileContent(any(SingleScenarioRunner.class), any(SingleScenario.class))).thenReturn("RUNNER_CONTENT");
 
         List<Path> pathList = new ArrayList<>();
         Path mockPath = mock(Path.class);