+++ CHANGELOG.md
@@ -6,6 +6,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 Back to [Readme](README.md).
 
+## [0.1.8] - 2018-03-30
+
+### Added
+
+- Source feature path is now included as a comment in generated features and runners
+- Configurable logging with 4 levels: _default_, _compact_, _minimal_, _off_
+- The number of generated scenarios per feature is shown in the default logs
+- The total number of generated features is shown in a summary line in all log levels except _off_
+
+### Changed
+
+- All dependencies are updated to their latest version
+
 ## [0.1.7] - 2018-03-02
 
 ### Changed
@@ -111,7 +124,8 @@ Back to [Readme](README.md).
 
 Initial project version on GitHub and Maven Central.
 
-[0.1.7]: https://github.com/trivago/cucable-plugin/compare/0.1.5...0.1.7
+[0.1.8]: https://github.com/trivago/cucable-plugin/compare/0.1.7...0.1.8
+[0.1.7]: https://github.com/trivago/cucable-plugin/compare/0.1.6...0.1.7
 [0.1.6]: https://github.com/trivago/cucable-plugin/compare/0.1.5...0.1.6
 [0.1.5]: https://github.com/trivago/cucable-plugin/compare/0.1.4...0.1.5
 [0.1.4]: https://github.com/trivago/cucable-plugin/compare/0.1.3...0.1.4